+++ datasetLoader/seq/Citeseer.py
@@ -73,7 +73,7 @@ class CiteseerGraph(GraphDataset):
     def process(self) -> None:
         X = torch.tensor(np.load(os.path.join(self.raw_dir, "citeseer_feat.npy")), dtype=torch.float)
         Y = torch.tensor(np.load(os.path.join(self.raw_dir, "citeseer_label.npy")), dtype=torch.long)
-        adj = torch.tensor(np.load(os.path.join(self.raw_dir, "citeseer_adj.npy")), dtype=torch.long)
+        adj = torch.tensor(np.load(os.path.join(self.raw_dir, "citeseer_adj.npy")), dtype=torch.float)
         
         adj_t = SparseTensor.from_dense(adj)
