+++ datasetLoader/seq/Cora.py
@@ -72,7 +72,7 @@ class CoraGraph(GraphDataset):
     def process(self) -> None:
         X = torch.tensor(np.load(os.path.join(self.raw_dir, "cora_feat.npy")), dtype=torch.float)
         Y = torch.tensor(np.load(os.path.join(self.raw_dir, "cora_label.npy")), dtype=torch.long)
-        adj = torch.tensor(np.load(os.path.join(self.raw_dir, "cora_adj.npy")), dtype=torch.long)
+        adj = torch.tensor(np.load(os.path.join(self.raw_dir, "cora_adj.npy")), dtype=torch.float)
         
         adj_t = SparseTensor.from_dense(adj)
         data = GraphData(x=X, y=Y, edge_index=adj_t)
