+++ datasetLoader/seq/CoraFull.py
@@ -67,7 +67,7 @@ class CoraFullGraph(GraphDataset):
     def process(self) -> None:
         X = torch.tensor(np.load(os.path.join(self.raw_dir, "corafull_feat.npy")), dtype=torch.float)
         Y = torch.tensor(np.load(os.path.join(self.raw_dir, "corafull_label.npy")), dtype=torch.long)
-        adj = torch.tensor(np.load(os.path.join(self.raw_dir, "corafull_adj.npy")), dtype=torch.long)
+        adj = torch.tensor(np.load(os.path.join(self.raw_dir, "corafull_adj.npy")), dtype=torch.float)
         
         adj_t = SparseTensor.from_dense(adj)
